+++ attributes/nameserver.rb
@@ -1,6 +1,0 @@
-default['cookbook-openshift3']['nameserver']['search'] = node['domain']
-default['cookbook-openshift3']['nameserver']['domain'] = nil
-default['cookbook-openshift3']['nameserver']['nameservers'] = []
-default['cookbook-openshift3']['nameserver']['key_algorithm'] = nil
-default['cookbook-openshift3']['nameserver']['key_name'] = nil
-default['cookbook-openshift3']['nameserver']['key_secret'] = nil